+++ src/http.rs
@@ -1,6 +1,6 @@
 use std::time::{Duration, Instant};
 
-use reqwest::header::HeaderValue;
+use reqwest::header::{HeaderMap, HeaderName, HeaderValue};
 use reqwest::{header, ClientBuilder};
 use url::Url;
 
@@ -33,15 +33,30 @@ impl std::fmt::Display for IpVersion {
 /// HTTP GET Request
 ///
 /// FROM provides an email address for the target host to be contacted in case of problems.
-/// See [http From header](https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/From)
-pub async fn get(url: &str, from: &str, accept_invalid_certs: bool) -> anyhow::Result<Response> {
+/// See [HTTP From header](https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/From)
+pub async fn get(
+    url: &str,
+    additional_headers: &Vec<String>,
+    from: &str,
+    accept_invalid_certs: bool,
+) -> anyhow::Result<Response> {
+    let mut headers = HeaderMap::new();
+    headers.insert(header::FROM, HeaderValue::from_str(from)?);
+    for entry in additional_headers {
+        let (k, v) = entry.split_once(": ").ok_or_else(|| {
+            anyhow::anyhow!("does not contain ': ' to separate header key/value: {entry}")
+        })?;
+        let name: HeaderName = k.parse()?;
+        headers.insert(name, v.parse()?);
+    }
+
     let request = ClientBuilder::new()
         .danger_accept_invalid_certs(accept_invalid_certs)
         .timeout(Duration::from_secs(30))
         .user_agent(USER_AGENT)
         .build()?
         .get(url)
-        .header(header::FROM, HeaderValue::from_str(from)?);
+        .headers(headers);
 
     let start = Instant::now();
     let response = request.send().await?.error_for_status()?;